64-bit: fix the pointer truncations and side-band gaps behind a Win64 crash
A downstream fork's Win64 build crashed loading a map that has no AAS
file. The zipped-AAS fallback passed (int)Path to getcwd, which keeps
only the low half of a stack address above 4 GB, so the CRT wrote
through a sign-extended kernel address. The same search passed
(int)Destination to the unzip helper, whose zip-name parameter is now a
real char pointer. Both casts were IDA's typing, not 1999 code: on
32-bit the int and the pointer are the same dword.

The report listed two more instances, and one more turned up alongside:

- BotGetMessageTeamGoal kept the waypoint it found in an int and copied
  the goal from 4 bytes past it. On 64-bit the pointer lost its high
  half, and the goal sits at 8, behind the name pointer. It now uses
  Quake III's own text, a typed waypoint and &cp->goal.
- BotShutdownClient freed the raw character slot, which is always 0 on
  64-bit because the real pointer lives in the side-band, so every bot
  shutdown leaked its character. It now frees BotCharacter(bs) and
  clears the side-band entry.
- BotMoveClient copied the bot state to the new slot but left all ten
  side-band entries at the old index. The moved bot ran with no
  character or weapon state, and the next bot set up in the old slot
  would have reused, and later double-freed, its weapon state. It now
  moves each entry and clears the source, as the 32-bit memset does.

The aarch64 binary scanners never see Windows-only code, so the
pointer-width warning sweep was re-run with the Win64 cross compiler.
It found nothing else live.

Verified: all six touched functions still match both oracles, and full
audits of the touched files are unchanged. Native, Win32 and Win64
builds are clean, and make verify passes. Under gdb, removing a bot now
frees a real character pointer, and on a real bot move the moved bot's
first AI frame found every side-band entry in its new slot.

Source-commit: eeab4b9f4525ace0190a7773eb9b5e3452e194c0

+++ botlib/be_aas_main.c
@@ -438,7 +438,7 @@ int __cdecl sub_1000E430(char *Source)
     strncat(Destination, v2, 144 - strlen(Destination));
     AppendPathSeperator(Destination, 144);
   }
-  getcwd_locked((int)Path, 144);
+  getcwd_locked(Path, 144);
   _chdir(Destination);
   if ( v2 )
     strncpy(dirs[0], v2, 0x90u);
@@ -462,7 +462,7 @@ int __cdecl sub_1000E430(char *Source)
       if ( !_access(Destination, 4) )
       {
         Log_Write("searching %s in %s", ArgList, Destination); /* "searching %s in %s" */
-        if ( sub_10041240((int)Destination, ArgList, 0) )
+        if ( sub_10041240(Destination, ArgList, 0) )
         {
           v7 = AAS_LoadAASFile(ArgList, 0, 0);
           errno = v7;
